refactor: #50 게임 시작 시 보스의 체력을 `인원수 * 기본값`으로 초기화하도록 로직 추가
